requirements.txt
pytest==5.3.5
requests==2.23.0
sphinx==2.4.3
sphinx-rtd-theme==0.4.3
python-dateutil==2.8.1